refactor: split redis_backend/jobs.rs into submodule
Move 1199-line jobs.rs into a folder module split by concern:
enqueue, dequeue, state, query, maintenance, errors, helpers.
Each submodule holds a partial impl block. No behavior changes.
